+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [PEP 440 Versioning](https://peps.python.org/pep-044
 - **Breaking:** Refactor the `ProjectsFinder` API and its associated implementation to improve performance and
   flexibility. <!-- TG-2228 -->
 - Sanitize `EnvironmentStatusResponse` API from `admin/environment/status` endpoint. <!-- TG-2183 -->
+- Package Version is now dynamic and stored in attribute `youwol.__version__` <!-- TG-2184 -->
 
 ### Fixed